Migrate `ComponentMetadataRulesIntegrationTest` to cross-repository test fixtures
